+++ vllm/model_executor/layers/fused_moe/experts/trtllm_fp8_moe.py
@@ -19,7 +19,7 @@ from vllm.model_executor.layers.fused_moe.utils import (
     fi_moe_largest_bucket,
     trtllm_moe_pack_topk_ids_weights,
 )
-from vllm.model_executor.layers.quantization.utils.flashinfer_utils import (
+from vllm.model_executor.layers.quantization.utils.flashinfer_moe import (
     activation_to_flashinfer_int,
 )
 from vllm.model_executor.layers.quantization.utils.quant_utils import (
